requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0
python-dotenv==0.21.0
whitenoise==6.4.0
django-cleanup==6.0.0
Pillow==9.2.0